Move and update test cases for Trie Add documentation for highjacking of tab key in CommandBox.java
